update: 各endpoint内でモジュール分けされていたのを削除
